Create BookList_View in the database and use decimal instead of double through out the entire solution

+++ Bulky.Models/TbBook.cs
@@ -12,13 +12,13 @@
 
         public string Author { get; set; } = null!;
 
-        public double ListPrice { get; set; }
+        public decimal ListPrice { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public double Price50 { get; set; }
+        public decimal Price50 { get; set; }
 
-        public double Price100 { get; set; }
+        public decimal Price100 { get; set; }
 
         public int CategoryId { get; set; }
         public TbCategory Category { get; set; } = null!;
